+++ widget-showcase/src/index.html
@@ -6,6 +6,5 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 </head>
 <body>
-	<showcase></showcase>
 </body>
 </html>
